Various improvements

+++ QuantumGateLib/Common/Random.cpp
@@ -22,8 +22,8 @@ namespace QuantumGate::Implementation
 		GetRngEngine().CheckSeed64(count);
 
 		// How many 64bit integers do we need for the number of requested bytes?
-		const auto num = static_cast<unsigned int>(std::ceil(static_cast<float>(count) /
-															 static_cast<float>(sizeof(UInt64))));
+		const auto num = static_cast<UInt64>(std::ceil(static_cast<double>(count) /
+													   static_cast<double>(sizeof(UInt64))));
 
 		assert(count <= (num * sizeof(UInt64)));
 
@@ -36,7 +36,7 @@ namespace QuantumGate::Implementation
 		{
 			return (GetRngEngine().Rng64)();
 		});
-		
+
 		bytes.Resize(count);
 
 		return bytes;
